GLOBAL/backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
python-dotenv>=1.0.1
duckdb>=1.1.0
pandas>=2.2.3
openpyxl>=3.1.5
openai>=1.54.0
ddgs>=9.0.0
langchain>=0.3.0
langchain-community>=0.3.0
langchain-text-splitters>=0.3.0
langchain-chroma>=0.1.0
chromadb>=0.5.0
langchain-google-genai>=2.0.0
